+++ solo_server/commands/robots/lerobot/__init__.py
@@ -1,0 +1,5 @@
+"""
+LeRobot utilities package for Solo Server
+Handles port detection, calibration, teleoperation setup, and data recording
+"""
+
