+++ solo_server/commands/robots/lerobot/utils/__init__.py
@@ -1,0 +1,3 @@
+"""
+LeRobot utilities subpackage
+"""
